+++ NetworkPerformanceSystem/NetworkPerformanceSystem.cs
@@ -32,7 +32,7 @@ namespace NetworkPerformanceSystem
     {
         public const string PluginGUID = "MidnightsFX.NetworkPerformanceSystem";
         public const string PluginName = "NetworkPerformanceSystem";
-        public const string PluginVersion = "1.2.1";
+        public const string PluginVersion = "1.3.0";
 
         internal static ManualLogSource Log;
         internal static Harmony HarmonyInstance;